// src/app/model/ctltlocation-model.ts
/**
 * Created by neng on 30/10/2559.
 */
export class CtltLocation {

    id: number;
    locCode: string;
    regId: number;
    locAbbrName: string;
    locName: string;
    locRepName: string;
    aumpCode: string;
    provCode: string;
    postcode: string;
    provName: string;
    tumbonName: string;
}